adding moveable drawer, and refactor persistent to use it
