+++ packages/mdc-drawer/movable/foundation.js
@@ -1,0 +1,250 @@
+/**
+ * Copyright 2016 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {MDCFoundation} from '@material/base';
+import {strings} from './constants';
+
+export class MDCMovableDrawerFoundation extends MDCFoundation {
+  static get strings() {
+    return strings;
+  }
+
+  static get defaultAdapter() {
+    return {
+      addClass: (/* className: string */) => {},
+      removeClass: (/* className: string */) => {},
+      hasClass: (/* className: string */) => {},
+      hasNecessaryDom: () => /* boolean */ false,
+      registerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
+      deregisterInteractionHandler: (/* evt: string, handler: EventListener */) => {},
+      registerDrawerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
+      deregisterDrawerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
+      registerTransitionEndHandler: (/* handler: EventListener */) => {},
+      deregisterTransitionEndHandler: (/* handler: EventListener */) => {},
+      registerDocumentKeydownHandler: (/* handler: EventListener */) => {},
+      deregisterDocumentKeydownHandler: (/* handler: EventListener */) => {},
+      setTranslateX: (/* value: number | null */) => {},
+      getFocusableElements: () => /* NodeList */ {},
+      saveElementTabState: (/* el: Element */) => {},
+      restoreElementTabState: (/* el: Element */) => {},
+      makeElementUntabbable: (/* el: Element */) => {},
+      isRtl: () => /* boolean */ false,
+      getDrawerWidth: () => /* number */ 0,
+      isDrawer: (/* el: Element */) => /* boolean */ false,
+    };
+  }
+
+  constructor(adapter) {
+    super(Object.assign(MDCMovableDrawerFoundation.defaultAdapter, adapter));
+
+    this.transitionEndHandler_ = (ev) => {
+      if (this.adapter_.isDrawer(ev.target)) {
+        this.adapter_.removeClass(this.animatingCssClass);
+        this.adapter_.deregisterTransitionEndHandler(this.transitionEndHandler_);
+      }
+    };
+
+    this.inert_ = false;
+
+    this.drawerClickHandler_ = (evt) => evt.stopPropagation();
+    this.componentTouchStartHandler_ = (evt) => this.handleTouchStart_(evt);
+    this.componentTouchMoveHandler_ = (evt) => this.handleTouchMove_(evt);
+    this.componentTouchEndHandler_ = (evt) => this.handleTouchEnd_(evt);
+    this.documentKeydownHandler_ = (evt) => {
+      if (evt.key && evt.key === 'Escape' || evt.keyCode === 27) {
+        this.close();
+      }
+    };
+  }
+
+  get rootCssClass() {
+    // Classes extending MDCMovableDrawerFoundation should implement this method to return a string
+    // CSS class for the drawer. e.g. 'mdc-temporary-drawer'
+    return "";
+  }
+
+  get animatingCssClass() {
+    // Classes extending MDCMovableDrawerFoundation should implement this method to return a string
+    // CSS class for when the drawer is animating. e.g. 'mdc-temporary-drawer--animating'
+    return "";
+  }
+
+  get openCssClass() {
+    // Classes extending MDCMovableDrawerFoundation should implement this method to return a string
+    // CSS class for when the drawer is animating. e.g. 'mdc-temporary-drawer--open'
+    return "";
+  }
+
+  init() {
+    const ROOT = this.rootCssClass;
+    const OPEN = this.openCssClass;
+
+    if (!this.adapter_.hasClass(ROOT)) {
+      throw new Error(`${ROOT} class required in root element.`);
+    }
+
+    if (!this.adapter_.hasNecessaryDom()) {
+      throw new Error(`Required DOM nodes missing in ${ROOT} component.`);
+    }
+
+    if (this.adapter_.hasClass(OPEN)) {
+      this.isOpen_ = true;
+    } else {
+      this.detabinate_();
+      this.isOpen_ = false;
+    }
+
+    this.adapter_.registerDrawerInteractionHandler('click', this.drawerClickHandler_);
+    this.adapter_.registerDrawerInteractionHandler('touchstart', this.componentTouchStartHandler_);
+    this.adapter_.registerInteractionHandler('touchmove', this.componentTouchMoveHandler_);
+    this.adapter_.registerInteractionHandler('touchend', this.componentTouchEndHandler_);
+  }
+
+  destroy() {
+    this.adapter_.deregisterDrawerInteractionHandler('click', this.drawerClickHandler_);
+    this.adapter_.deregisterDrawerInteractionHandler('touchstart', this.componentTouchStartHandler_);
+    this.adapter_.deregisterInteractionHandler('touchmove', this.componentTouchMoveHandler_);
+    this.adapter_.deregisterInteractionHandler('touchend', this.componentTouchEndHandler_);
+    // Deregister the document keydown handler just in case the component is destroyed while the menu is open.
+    this.adapter_.deregisterDocumentKeydownHandler(this.documentKeydownHandler_);
+  }
+
+  open() {
+    this.adapter_.registerTransitionEndHandler(this.transitionEndHandler_);
+    this.adapter_.registerDocumentKeydownHandler(this.documentKeydownHandler_);
+    this.adapter_.addClass(this.animatingCssClass);
+    this.adapter_.addClass(this.openCssClass);
+    this.retabinate_();
+    this.isOpen_ = true;
+  }
+
+  close() {
+    this.adapter_.deregisterDocumentKeydownHandler(this.documentKeydownHandler_);
+    this.adapter_.registerTransitionEndHandler(this.transitionEndHandler_);
+    this.adapter_.addClass(this.animatingCssClass);
+    this.adapter_.removeClass(this.openCssClass);
+    this.detabinate_();
+    this.isOpen_ = false;
+  }
+
+  isOpen() {
+    return this.isOpen_;
+  }
+
+  /**
+   *  Render all children of the drawer inert when it's closed.
+   */
+  detabinate_() {
+    if (this.inert_) {
+      return;
+    }
+
+    const elements = this.adapter_.getFocusableElements();
+    if (elements) {
+      for (let i = 0; i < elements.length; i++) {
+        this.adapter_.saveElementTabState(elements[i]);
+        this.adapter_.makeElementUntabbable(elements[i]);
+      }
+    }
+
+    this.inert_ = true;
+  }
+
+  /**
+   *  Make all children of the drawer tabbable again when it's open.
+   */
+  retabinate_() {
+    if (!this.inert_) {
+      return;
+    }
+
+    const elements = this.adapter_.getFocusableElements();
+    if (elements) {
+      for (let i = 0; i < elements.length; i++) {
+        this.adapter_.restoreElementTabState(elements[i]);
+      }
+    }
+
+    this.inert_ = false;
+  }
+
+  handleTouchStart_(evt) {
+    if (!this.adapter_.hasClass(this.openCssClass)) {
+      return;
+    }
+    if (evt.pointerType && evt.pointerType !== 'touch') {
+      return;
+    }
+
+    this.direction_ = this.adapter_.isRtl() ? -1 : 1;
+    this.drawerWidth_ = this.adapter_.getDrawerWidth();
+    this.startX_ = evt.touches ? evt.touches[0].pageX : evt.pageX;
+    this.currentX_ = this.startX_;
+    this.touchingSideNav_ = true;
+
+    requestAnimationFrame(this.updateDrawer_.bind(this));
+  }
+
+  handleTouchMove_(evt) {
+    if (evt.pointerType && evt.pointerType !== 'touch') {
+      return;
+    }
+
+    this.currentX_ = evt.touches ? evt.touches[0].pageX : evt.pageX;
+  }
+
+  handleTouchEnd_(evt) {
+    if (evt.pointerType && evt.pointerType !== 'touch') {
+      return;
+    }
+
+    this.touchingSideNav_ = false;
+    this.adapter_.setTranslateX(null);
+
+    let newPos = null;
+    if (this.direction_ === 1) {
+      newPos = Math.min(0, this.currentX_ - this.startX_);
+    } else {
+      newPos = Math.max(0, this.currentX_ - this.startX_);
+    }
+
+    // Did the user close the drawer by more than 50%?
+    if (Math.abs(newPos / this.drawerWidth_) >= 0.5) {
+      this.close();
+    } else {
+      // Triggering an open here means we'll get a nice animation back to the fully open state.
+      this.open();
+    }
+  }
+
+  updateDrawer_() {
+    if (!this.touchingSideNav_) {
+      return;
+    }
+
+    requestAnimationFrame(this.updateDrawer_.bind(this));
+
+    let newPos = null;
+
+    if (this.direction_ === 1) {
+      newPos = Math.min(0, this.currentX_ - this.startX_);
+    } else {
+      newPos = Math.max(0, this.currentX_ - this.startX_);
+    }
+
+    this.adapter_.setTranslateX(newPos);
+  }
+}
