+++ packages/mdc-drawer/movable/index.js
@@ -14,8 +14,4 @@
  * limitations under the License.
  */
 
-import * as util from './util';
-export {MDCMovableDrawerFoundation} from './movable';
-export {MDCTemporaryDrawer, MDCTemporaryDrawerFoundation} from './temporary';
-export {MDCPersistentDrawer, MDCPersistentDrawerFoundation} from './persistent';
-export {util};
+export {MDCMovableDrawerFoundation} from './foundation';
