+++ packages/mdc-drawer/persistent/constants.js
@@ -21,8 +21,3 @@ export const cssClasses = {
   OPEN: `${ROOT}--open`,
   ANIMATING: `${ROOT}--animating`,
 };
-
-export const strings = {
-  FOCUSABLE_ELEMENTS: 'a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), ' +
-      'button:not([disabled]), iframe, object, embed, [tabindex], [contenteditable]',
-};
