+++ packages/mdc-drawer/persistent/foundation.js
@@ -14,222 +14,23 @@
  * limitations under the License.
  */
 
-import {MDCFoundation} from '@material/base';
-import {cssClasses, strings} from './constants';
+import {MDCMovableDrawerFoundation} from '../movable';
+import {cssClasses} from './constants';
 
-export default class MDCPersistentDrawerFoundation extends MDCFoundation {
+export default class MDCPersistentDrawerFoundation extends MDCMovableDrawerFoundation {
   static get cssClasses() {
     return cssClasses;
   }
 
-  static get strings() {
-    return strings;
+  get rootCssClass() {
+    return MDCPersistentDrawerFoundation.cssClasses.ROOT;
   }
 
-  static get defaultAdapter() {
-    return {
-      addClass: (/* className: string */) => {},
-      removeClass: (/* className: string */) => {},
-      hasClass: (/* className: string */) => {},
-      hasNecessaryDom: () => /* boolean */ false,
-      registerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
-      deregisterInteractionHandler: (/* evt: string, handler: EventListener */) => {},
-      registerDrawerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
-      deregisterDrawerInteractionHandler: (/* evt: string, handler: EventListener */) => {},
-      registerTransitionEndHandler: (/* handler: EventListener */) => {},
-      deregisterTransitionEndHandler: (/* handler: EventListener */) => {},
-      registerDocumentKeydownHandler: (/* handler: EventListener */) => {},
-      deregisterDocumentKeydownHandler: (/* handler: EventListener */) => {},
-      setTranslateX: (/* value: number | null */) => {},
-      getFocusableElements: () => /* NodeList */ {},
-      saveElementTabState: (/* el: Element */) => {},
-      restoreElementTabState: (/* el: Element */) => {},
-      makeElementUntabbable: (/* el: Element */) => {},
-      isRtl: () => /* boolean */ false,
-      getDrawerWidth: () => /* number */ 0,
-      isDrawer: (/* el: Element */) => /* boolean */ false,
-    };
+  get animatingCssClass() {
+    return MDCPersistentDrawerFoundation.cssClasses.ANIMATING;
   }
 
-  constructor(adapter) {
-    super(Object.assign(MDCPersistentDrawerFoundation.defaultAdapter, adapter));
-
-    this.transitionEndHandler_ = (ev) => {
-      if (this.adapter_.isDrawer(ev.target)) {
-        this.adapter_.removeClass(MDCPersistentDrawerFoundation.cssClasses.ANIMATING);
-        this.adapter_.deregisterTransitionEndHandler(this.transitionEndHandler_);
-      }
-    };
-
-    this.inert_ = false;
-
-    this.drawerClickHandler_ = (evt) => evt.stopPropagation();
-    this.componentTouchStartHandler_ = (evt) => this.handleTouchStart_(evt);
-    this.componentTouchMoveHandler_ = (evt) => this.handleTouchMove_(evt);
-    this.componentTouchEndHandler_ = (evt) => this.handleTouchEnd_(evt);
-    this.documentKeydownHandler_ = (evt) => {
-      if (evt.key && evt.key === 'Escape' || evt.keyCode === 27) {
-        this.close();
-      }
-    };
-  }
-
-  init() {
-    const {ROOT, OPEN} = MDCPersistentDrawerFoundation.cssClasses;
-
-    if (!this.adapter_.hasClass(ROOT)) {
-      throw new Error(`${ROOT} class required in root element.`);
-    }
-
-    if (!this.adapter_.hasNecessaryDom()) {
-      throw new Error(`Required DOM nodes missing in ${ROOT} component.`);
-    }
-
-    if (this.adapter_.hasClass(OPEN)) {
-      this.isOpen_ = true;
-    } else {
-      this.detabinate_();
-      this.isOpen_ = false;
-    }
-
-    this.adapter_.registerDrawerInteractionHandler('click', this.drawerClickHandler_);
-    this.adapter_.registerDrawerInteractionHandler('touchstart', this.componentTouchStartHandler_);
-    this.adapter_.registerInteractionHandler('touchmove', this.componentTouchMoveHandler_);
-    this.adapter_.registerInteractionHandler('touchend', this.componentTouchEndHandler_);
-  }
-
-  destroy() {
-    this.adapter_.deregisterDrawerInteractionHandler('click', this.drawerClickHandler_);
-    this.adapter_.deregisterDrawerInteractionHandler('touchstart', this.componentTouchStartHandler_);
-    this.adapter_.deregisterInteractionHandler('touchmove', this.componentTouchMoveHandler_);
-    this.adapter_.deregisterInteractionHandler('touchend', this.componentTouchEndHandler_);
-    // Deregister the document keydown handler just in case the component is destroyed while the menu is open.
-    this.adapter_.deregisterDocumentKeydownHandler(this.documentKeydownHandler_);
-  }
-
-  open() {
-    this.adapter_.registerTransitionEndHandler(this.transitionEndHandler_);
-    this.adapter_.registerDocumentKeydownHandler(this.documentKeydownHandler_);
-    this.adapter_.addClass(MDCPersistentDrawerFoundation.cssClasses.ANIMATING);
-    this.adapter_.addClass(MDCPersistentDrawerFoundation.cssClasses.OPEN);
-    this.retabinate_();
-    this.isOpen_ = true;
-  }
-
-  close() {
-    this.adapter_.deregisterDocumentKeydownHandler(this.documentKeydownHandler_);
-    this.adapter_.registerTransitionEndHandler(this.transitionEndHandler_);
-    this.adapter_.addClass(MDCPersistentDrawerFoundation.cssClasses.ANIMATING);
-    this.adapter_.removeClass(MDCPersistentDrawerFoundation.cssClasses.OPEN);
-    this.detabinate_();
-    this.isOpen_ = false;
-  }
-
-  isOpen() {
-    return this.isOpen_;
-  }
-
-  /**
-   *  Render all children of the drawer inert when it's closed.
-   */
-  detabinate_() {
-    if (this.inert_) {
-      return;
-    }
-
-    const elements = this.adapter_.getFocusableElements();
-    if (elements) {
-      for (let i = 0; i < elements.length; i++) {
-        this.adapter_.saveElementTabState(elements[i]);
-        this.adapter_.makeElementUntabbable(elements[i]);
-      }
-    }
-
-    this.inert_ = true;
-  }
-
-  /**
-   *  Make all children of the drawer tabbable again when it's open.
-   */
-  retabinate_() {
-    if (!this.inert_) {
-      return;
-    }
-
-    const elements = this.adapter_.getFocusableElements();
-    if (elements) {
-      for (let i = 0; i < elements.length; i++) {
-        this.adapter_.restoreElementTabState(elements[i]);
-      }
-    }
-
-    this.inert_ = false;
-  }
-
-  handleTouchStart_(evt) {
-    if (!this.adapter_.hasClass(MDCPersistentDrawerFoundation.cssClasses.OPEN)) {
-      return;
-    }
-    if (evt.pointerType && evt.pointerType !== 'touch') {
-      return;
-    }
-
-    this.direction_ = this.adapter_.isRtl() ? -1 : 1;
-    this.drawerWidth_ = this.adapter_.getDrawerWidth();
-    this.startX_ = evt.touches ? evt.touches[0].pageX : evt.pageX;
-    this.currentX_ = this.startX_;
-    this.touchingSideNav_ = true;
-
-    requestAnimationFrame(this.updateDrawer_.bind(this));
-  }
-
-  handleTouchMove_(evt) {
-    if (evt.pointerType && evt.pointerType !== 'touch') {
-      return;
-    }
-
-    this.currentX_ = evt.touches ? evt.touches[0].pageX : evt.pageX;
-  }
-
-  handleTouchEnd_(evt) {
-    if (evt.pointerType && evt.pointerType !== 'touch') {
-      return;
-    }
-
-    this.touchingSideNav_ = false;
-    this.adapter_.setTranslateX(null);
-
-    let newPos = null;
-    if (this.direction_ === 1) {
-      newPos = Math.min(0, this.currentX_ - this.startX_);
-    } else {
-      newPos = Math.max(0, this.currentX_ - this.startX_);
-    }
-
-    // Did the user close the drawer by more than 50%?
-    if (Math.abs(newPos / this.drawerWidth_) >= 0.5) {
-      this.close();
-    } else {
-      // Triggering an open here means we'll get a nice animation back to the fully open state.
-      this.open();
-    }
-  }
-
-  updateDrawer_() {
-    if (!this.touchingSideNav_) {
-      return;
-    }
-
-    requestAnimationFrame(this.updateDrawer_.bind(this));
-
-    let newPos = null;
-
-    if (this.direction_ === 1) {
-      newPos = Math.min(0, this.currentX_ - this.startX_);
-    } else {
-      newPos = Math.max(0, this.currentX_ - this.startX_);
-    }
-
-    this.adapter_.setTranslateX(newPos);
+  get openCssClass() {
+    return MDCPersistentDrawerFoundation.cssClasses.OPEN;
   }
 }
